requirements.txt
altair==5.3.0
annotated-types==0.7.0
anthropic==0.31.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.2.0
autopep8==2.3.1
backports.tarfile==1.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
Brotli==1.1.0
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
dacite==1.8.1
decorator==4.4.2
Deprecated==1.2.14
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
docopt==0.6.2
docutils==0.21.2
exceptiongroup==1.2.1
fastapi==0.103.2
feedparser==6.0.11
ffmpy==0.3.2
filelock==3.15.4
FLAML==2.1.2
fsspec==2024.6.1
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
importlib_metadata==8.0.0
jaconv==0.3.4
jaraco.classes==3.4.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jeepney==0.8.0
Jinja2==3.1.4
jiter==0.5.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keyring==25.2.1
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.3.0
moviepy==1.0.3
mpmath==1.3.0
music-tag==0.4.3
mutagen==1.47.0
networkx==3.3
nh3==0.2.18
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
openai==1.35.13
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pkginfo==1.10.0
platformdirs==4.2.2
proglog==0.1.10
protobuf==5.27.2
psutil==6.0.0
pyarrow==16.1.0
pyautogen==0.2.32
pycodestyle==2.12.0
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
pykakasi==2.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytube==15.0.0
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.9.4
readme_renderer==44.0
redis==5.0.7
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
rpds-py==0.19.0
savify==2.3.4
SecretStorage==3.3.3
sgmllib3k==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soundcloud-v2==1.5.3
soupsieve==2.5
spotdl==4.2.5
spotipy==2.24.0
starlette==0.27.0
stcli==0.2.1
streamlit==1.36.0
sympy==1.13.0
syncedlyrics==0.7.0
tenacity==8.5.0
termcolor==2.3.0
text-unidecode==1.3
tiktoken==0.7.0
tldextract==5.1.2
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.4
triton==2.3.1
twine==5.1.1
typing_extensions==4.12.2
tzdata==2024.1
Unidecode==1.3.8
urllib3==2.2.2
uvicorn==0.23.2
validators==0.32.0
watchdog==4.0.1
websockets==12.0
wrapt==1.16.0
yaspin==3.0.2
youtube-dl==2021.12.17
yt-dlp==2024.7.9
ytmusicapi==1.7.4
zipp==3.19.2
